split off MoR component tests to different script
